* tag '0.12': (949 commits)
  ENH more robust transformer testing.... don't ask why that came up
  Use pinvh wherever it helps in the codebase.
  Cloned @jakevdp's pinvh tests
  @jakevdp's version of pinvh speed up symmetric_pinv
  Add comments on optimized precision computations.
  Vectorize singular value inversion
  Compute pseudoinverse using eigendecomposition
  We already have the inverse at that step
  DOC added link to 0.11 docs to support page.
  DOC add people and commits do whatsnew
  MISC changed version number for release, change maintainer to myself
  COSMIT typo, thanks @ogrisel
  COSMIT pep8
  small tweaks
  typos and alex`s review changes
  docstring changes
  not the problem afterall - switch back
  changed includes back - change broke JENKINS build
  docstring fixes
  docstring change
  ...

+++ sklearn/svm/liblinear.pyx
@@ -10,8 +10,8 @@ cimport liblinear
 
 
 def train_wrap ( np.ndarray[np.float64_t, ndim=2, mode='c'] X,
-                 np.ndarray[np.int32_t, ndim=1, mode='c'] Y, int
-                 solver_type, double eps, double bias, double C, 
+                 np.ndarray[np.float64_t, ndim=1, mode='c'] Y,
+                 int solver_type, double eps, double bias, double C, 
                  np.ndarray[np.int32_t, ndim=1] weight_label,
                  np.ndarray[np.float64_t, ndim=1] weight):
     """
@@ -66,8 +66,8 @@ cdef _csr_train_wrap(np.int32_t n_features,
                  np.ndarray[np.float64_t, ndim=1, mode='c'] X_values,
                  np.ndarray[np.int32_t,   ndim=1, mode='c'] X_indices,
                  np.ndarray[np.int32_t,   ndim=1, mode='c'] X_indptr,
-                 np.ndarray[np.int32_t, ndim=1, mode='c'] Y, int
-                 solver_type, double eps, double bias, double C,
+                 np.ndarray[np.float64_t, ndim=1, mode='c'] Y,
+                 int solver_type, double eps, double bias, double C,
                  np.ndarray[np.int32_t, ndim=1] weight_label,
                  np.ndarray[np.float64_t, ndim=1] weight):
     cdef parameter *param
